Revert "lint: remove unnecessary type conversions"
This reverts commit c7451c12a1b1bf1143d23710d6dd294d958b1fb3.

+++ client/go/internal/vespa/document/http.go
@@ -352,6 +352,6 @@ func (c *Client) resultWithResponse(resp *http.Response, sentBytes int, result R
 	}
 	result.Latency = elapsed
 	result.BytesSent = int64(sentBytes)
-	result.BytesRecv = written
+	result.BytesRecv = int64(written)
 	return result
 }
